+++ recipes/excluder.rb
@@ -16,6 +16,5 @@ ose_major_version = node['is_apaas_openshift_cookbook']['deploy_containerized'] 
   execute "Enable atomic-openshift-#{pkg}" do
     command "atomic-openshift-#{pkg} disable"
     not_if { ose_major_version.split('.')[1].to_i < 5 && node['is_apaas_openshift_cookbook']['openshift_deployment_type'] != 'enterprise' }
-    not_if { node['is_apaas_openshift_cookbook']['upgrade'] }
   end
 end