+++ recipes/upgrade_control_plane36.rb
@@ -90,7 +90,7 @@ if ::File.file?(node['is_apaas_openshift_cookbook']['control_upgrade_flag'])
     node.force_override['is_apaas_openshift_cookbook']['etcd_migrated'] = false unless config_options['kubernetesMasterConfig']['apiServerArguments'].key?('storage-backend')
 
     include_recipe 'is_apaas_openshift_cookbook::master'
-    include_recipe 'is_apaas_openshift_cookbook::excluder'
+    include_recipe 'is_apaas_openshift_cookbook::excluder' unless is_node_server
 
     log 'Restart Master services' do
       level :info