+++ recipes/upgrade_control_plane37_part2.rb
@@ -63,7 +63,7 @@ if is_master_server
   end
 
   include_recipe 'is_apaas_openshift_cookbook::master'
-  include_recipe 'is_apaas_openshift_cookbook::excluder'
+  include_recipe 'is_apaas_openshift_cookbook::excluder' unless is_node_server
 
   log 'Restart Master services' do
     level :info